Set log field to all possible integer size.
Tested. Use this to generate log processing program for COREMU.

+++ mem.h
@@ -22,10 +22,10 @@ extern "C" {
 #define OBJ_SIZE 8
 #define NOBJS 10
 
-typedef int version_t;
-typedef int objid_t;
+typedef long version_t;
+typedef int16_t objid_t;
 typedef int memop_t;
-typedef int tid_t;
+typedef int8_t tid_t;
 
 // WaitMemop is used when recording
 typedef struct {
